Fix config copy in installer

+++ install.sh
@@ -12,4 +12,21 @@ make build-python PYTHON=$(which python3)
 sudo make install-python PYTHON=$(which python3)
 
 cd ../..
-cp config.json.example config.json
+
+if [ -n "$1" ]; then
+    echo -e "\nSkipping config.json file creation"
+else
+    echo -e "\nYou'll need a config.json file to customize your settings. If you are updating"
+    echo "from an older version and you were required to run this install script again or"
+    echo "this is a fresh install, it's recommended we make a fresh one right now."
+    echo "This will create a brand new 'config.json' file with default values so edit this"
+    echo -e "file with your own settings.\n"
+    read -p "Would you like to do this now? [Y/n] " answer
+    if [ "$answer" != "${answer#[Yy]}" ] ;then
+        rm config.json
+        cp config.json.example config.json
+        echo -e "\nYou should now have a fresh config.json file you can customize with your own settings.\n"
+    else
+        echo -e "\nIf you do not have a config.json, you can manually copy the config.json.example to config.json to customize settings.\n"
+    fi
+fi
